Code refractoring for order of operations

+++ timesheet_backend/src/main/java/com/g4l/timesheet_backend/services/implementations/ClientServiceImpl.java
@@ -13,6 +13,7 @@ import com.g4l.timesheet_backend.models.responses.ClientTeamResponse;
 import com.g4l.timesheet_backend.repositories.ClientRepository;
 import com.g4l.timesheet_backend.repositories.ClientTeamRepository;
 import com.g4l.timesheet_backend.services.interfaces.ClientService;
+import com.g4l.timesheet_backend.services.interfaces.ConsultantService;
 import com.g4l.timesheet_backend.services.interfaces.ManagerService;
 import com.g4l.timesheet_backend.utils.SequenceGenerator;
 import com.g4l.timesheet_backend.utils.mappers.models.ClientMapper;
@@ -23,13 +24,20 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class ClientServiceImpl implements ClientService {
 
+    // TODO: Figure out which level to map user based responses
+    // TODO: Figure out whether to use userId or any idenfier for user methods
+
     private final ClientRepository clientRepository;
     private final ClientTeamRepository clientTeamRepository;
     private final ClientMapper clientMapper;
     private final ManagerService managerService;
+    private final ConsultantService consultantService;
 
     @Override
     public Object createClient(String clientName) {
+        if (getClientByName(clientName) != null)
+            return null;
+
         Client client = new Client();
 
         client.setId(SequenceGenerator.generateSequence(SequenceType.CLIENT_ID));
@@ -47,6 +55,9 @@ public class ClientServiceImpl implements ClientService {
 
     @Override
     public Object updateClient(ClientRequest clientRequest) {
+        if (getClientById(clientRequest.getId()) == null)
+            return null;
+
         Client client = clientMapper.clientRequestToClient(clientRequest);
 
         client.setDateModified(LocalDateTime.now());
@@ -67,6 +78,9 @@ public class ClientServiceImpl implements ClientService {
 
     @Override
     public Object createClientTeam(ClientTeamRequest clientTeamRequest) {
+        if (getClientTeamByName(clientTeamRequest.getTeamName()) != null)
+            return null;
+
         ClientTeam clientTeam = clientMapper.clientTeamRequestToClientTeam(clientTeamRequest);
 
         clientTeam.setId(SequenceGenerator.generateSequence(SequenceType.TEAM_ID));
@@ -84,6 +98,9 @@ public class ClientServiceImpl implements ClientService {
 
     @Override
     public Object updateClientTeam(ClientTeamRequest clientTeamRequest) {
+        if (getClientTeamById(clientTeamRequest.getId()) == null)
+            return null;
+
         ClientTeam clientTeam = clientMapper.clientTeamRequestToClientTeam(clientTeamRequest);
 
         clientTeam.setDateModified(LocalDateTime.now());
@@ -104,6 +121,9 @@ public class ClientServiceImpl implements ClientService {
 
     @Override
     public Object deleteClient(@NonNull String clientId) {
+        if (getClientById(clientId) == null)
+            return null;
+
         try {
             clientRepository.deleteById(clientId);
             return "Client with userId " + clientId + " deleted";
@@ -114,6 +134,9 @@ public class ClientServiceImpl implements ClientService {
 
     @Override
     public Object deleteClientTeam(@NonNull String clientTeamId) {
+        if (getClientTeamById(clientTeamId) == null)
+            return null;
+
         try {
             clientTeamRepository.deleteById(clientTeamId);
             return "ClientTeam with userId " + clientTeamId + " deleted";
@@ -136,15 +159,34 @@ public class ClientServiceImpl implements ClientService {
 
     @Override
     public Object assignTeamToManager(String managerId, String teamId) {
-        if (getClientTeamById(teamId) == null) {
+        if (getClientTeamById(teamId) == null)
+            return null;
+
+        if (managerService.assignTeamToManager(managerId, teamId) == null) {
             return null;
         }
 
-        if(managerService.assignTeamToManager(managerId, teamId) == null) {
-            return null;
-        }   
-
-        return (Object) managerService.assignTeamToManager(managerId, teamId);
+        return "Manager with manager id " + managerId + " assigned to team with team id " + teamId;
     }
 
+    @Override
+    public Object assignConsultantToTeam(String consultantId, String teamId) {
+        if (getClientTeamById(teamId) == null)
+            return null;
+
+        if (consultantService.assignConsultantToClientTeam(consultantId, teamId) != null)
+            return null;
+
+        return "Consultant with consultant id " + consultantId + " assigned to team with team id " + teamId;
+    }
+
+    @Override
+    public Object getClientByName(String clientName) {
+        return clientRepository.findByClientName(clientName);
+    }
+
+    @Override
+    public Object getClientTeamByName(String clientTeamName) {
+        return clientTeamRepository.findByClientTeamName(clientTeamName);
+    }
 }
